+++ src/main/java/ru/vlsu/fitclub/repositories/TrainerSpecializationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.vlsu.fitclub.model.TrainerSpecialization;
 
 @Repository
-public interface TrainerSpecializationRepository extends CrudRepository<TrainerSpecialization, Long> {
+public interface TrainerSpecializationRepository extends CrudRepository<TrainerSpecialization, Integer> {
 }
